client & server: 文章详情页调试...

+++ client/src/components/header/index.vue
@@ -1,12 +1,12 @@
 <template>
   <div class="header-container">
-    <div class="header-container-left" @click="jumpPage('main')">BLOG</div>
+    <div class="header-container-left" @click="jumpPage('/')">BLOG</div>
     <div class="header-container-right">
       <template v-if="user.isUser">
-        <el-button type="info" :icon="Edit" link @click="jumpPage('articles')">创建新文章</el-button>
-        <el-button type="info" :icon="Setting" link>用户配置</el-button>
+        <el-button type="info" :icon="Edit" link @click="jumpPage('/articles')">创建新文章</el-button>
+        <el-button type="info" :icon="Setting" link @click="jumpPage('/userSetting')">用户配置</el-button>
 
-        <avatar :user="userAvatar"></avatar>
+        <avatar :user="userAvatar" @click="jumpPage('/user/' + user.userName)"></avatar>
 
         <el-button type="danger" link @click="handleLogout">登 出</el-button>
       </template>
@@ -22,10 +22,13 @@
 import { computed } from 'vue'
 import { useRouter, useRoute } from 'vue-router'
 import { useUserStore } from '@/stores/user'
-import { Edit, Setting, User, Plus } from '@element-plus/icons-vue'
 import { Status, useAccountStore } from '@/stores/account'
 
+import { ElMessageBox } from 'element-plus'
+import { Edit, Setting, User, Plus } from '@element-plus/icons-vue'
 import avatar from '@/components/avatar/index.vue'
+
+import type { RouteParamsRaw } from 'vue-router'
 
 /** Use of external methods */
 const user = useUserStore()
@@ -45,18 +48,26 @@ const userAvatar = computed(() => {
 
 /** Operation */
 const handleLogout = () => {
-  user.resetUser()
-  window.sessionStorage.clear()
-  jumpPage('main')
+  ElMessageBox.confirm('确认退出?', '提示', {
+    confirmButtonText: '确认',
+    cancelButtonText: '取消',
+    type: 'error'
+  }).then(() => {
+    user.resetUser()
+    window.sessionStorage.clear()
+    jumpPage('/')
+  })
 }
 const handleAccount = (status: Status) => {
   account.setStatus(status)
 
-  if (route.path !== 'account') {
-    jumpPage('account')
+  if (route.path !== '/account') {
+    jumpPage('/account')
   }
 }
-const jumpPage = (name: string) => router.push({ name })
+const jumpPage = (path: string, params?: RouteParamsRaw) => {
+  router.push({ path, params: params || {} })
+}
 </script>
 
 <style lang="scss" scoped>
